refactor: Move TUI package under internal/interactive namespace
