+++ internal/interactive/tui/cmdSelector.go
