+++ internal/interactive/tui/commandBot.go
