+++ internal/interactive/tui/commandMcp.go
