+++ internal/interactive/tui/commandMcpAdd.go
