+++ internal/interactive/tui/commandMcpRemove.go
