+++ internal/interactive/tui/commandMode.go
