+++ internal/interactive/tui/commandModel.go
