+++ internal/interactive/tui/commandModelAdd.go
