+++ internal/interactive/tui/commandNew.go
