+++ internal/interactive/tui/commandPlanner.go
