+++ internal/interactive/tui/commandReasoning.go
