+++ internal/interactive/tui/commandSessionModel.go
