+++ internal/interactive/tui/commandSwitch.go
