+++ internal/interactive/tui/commnadModelRemove.go
