+++ internal/interactive/tui/handlerCommand.go
