+++ internal/interactive/tui/handlerFetchReleased.go
