+++ internal/interactive/tui/handlerInputHistory.go
