+++ internal/interactive/tui/handlerLogFormat.go
