+++ internal/interactive/tui/handlerLogMode.go
