+++ internal/interactive/tui/helper.go
