+++ internal/interactive/tui/view.go
