+++ internal/interactive/tui/viewRender.go
